+++ src/main/java/com/kdh/domain/UserVo.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserVo {
-	private int user_idx;
+	private Long user_idx;
 	private String user_id;
 	private String user_pw;
 	private String user_name;
